Fix inc file path; Fix macros; Add inclusion guards
