+++ @AresModAchillesExpansion/addons/functions_f/cfgFunctions.hpp
@@ -1,16 +1,14 @@
-#include "\achilles\function_f\includes\macros.inc.sqf"
-
 class CfgFunctions
 {
 	class Achilles
 	{
 		project = "Achilles";
 		tag = "Achilles";
-
+		
 		class Module
 		{
-			DEF_FUNC_ACHIL_1(module, init)
-			DEF_FUNC_ACHIL_1(module, openDialog)
+			DEF_FUNC_ACHIL_1(module,init);
+			DEF_FUNC_ACHIL_1(module,openDialog);
 		};
 	};
 };
