+++ @AresModAchillesExpansion/addons/language_f/config.cpp
@@ -1,3 +1,5 @@
+#define ACHILLES_CONFIG_ENV
+
 class CfgPatches
 {
 	class Achilles_Language_F
