--- @AresModAchillesExpansion/addons/modules_f/cfgFactionClasses.hpp
@@ -1,9 +1,0 @@
-
-class CfgFactionClasses
-{
-	class NO_CATEGORY;
-	
-	class ace_zeus_Medical {displayName = "ACE Medical";};
-	class Achilles_fac_Buildings: NO_CATEGORY {displayName = "$STR_AMAE_BUILDINGS";};
-	class Achilles_fac_Objects: NO_CATEGORY {displayName = "$STR_AMAE_OBJECTS";};
-};
