+++ @AresModAchillesExpansion/addons/modules_f/cfgFunctions.hpp
@@ -1,23 +1,8 @@
-
 class CfgFunctions
 {
 	class Achilles // This bit will be prefixed when actually calling the function (e.g. "Achilles_fnc_...." )
 	{
-		#include "ACE\cfgFunctionsACE.hpp"
-		#include "Arsenal\cfgFunctionsArsenal.hpp"
-		#include "Behaviours\cfgFunctionsBehaviours.hpp"
-		#include "DevTools\cfgFunctionsDevTools.hpp"
-		#include "Environment\cfgFunctionsEnvironment.hpp"
-		#include "FireSupport\cfgFunctionsFireSupport.hpp"
-		#include "Buildings\cfgFunctionsBuildings.hpp"
-		#include "Effects\cfgFunctionsEffects.hpp"
-		#include "Equipment\cfgFunctionsEquipment.hpp"
-		#include "MissionFlow\cfgFunctionsMissionFlow.hpp"
-		#include "Objects\cfgFunctionsObjects.hpp"
-		#include "Player\cfgFunctionsPlayer.hpp"
-		#include "Reinforcements\cfgFunctionsReinforcements.hpp"
-		#include "Replacement\cfgFunctionsReplacement.hpp"
-		#include "Spawn\cfgFunctionsSpawn.hpp"
-		#include "Zeus\cfgFunctionsZeus.hpp"
+		project = "Achilles";
+		tag = "Achilles";
 	};
 };
