+++ @AresModAchillesExpansion/addons/modules_f_achilles/config.cpp
@@ -124,8 +124,8 @@ class CfgPatches
 			"A3_Modules_F_Curator",
 			"achilles_Language_F",
 			"achilles_Functions_F_Ares",
-			"achilles_Functions_F_achilles",
-			"achilles_Data_F_achilles",
+			"achilles_Functions_F_Achilles",
+			"achilles_Data_F_Achilles",
 			"achilles_Data_F_Ares"
 		};
 	};
